refactor: extract RunState agent graph identity
